make the main run function be exported for use elsewhere
